requirements.txt
Flask==0.11.1
Jinja2==2.8
Flask-login
Flask-sqlalchemy
Flask-admin
Flask-wtf
Flask-babel
Flask-BasicAuth
werkzeug
stripe
urllib3
pyopenssl
ndg-httpsclient
pyasn1
Flask-WhooshAlchemy
sqlalchemy-migrate==0.10.0
itsdangerous
pyOpenSSL==16.2.0
flask_mail